+++ src/GlobalPayments.Api/Entities/FraudRule.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace GlobalPayments.Api.Entities
+{
+    public class FraudRule
+    {
+        public string Key { get; set; }
+        public FraudFilterMode Mode { get; set; }
+    }
+}