+++ src/GlobalPayments.Api/Gateways/GpEcomConnector.cs
@@ -73,8 +73,8 @@ namespace GlobalPayments.Api.Gateways {
                 var card = builder.PaymentMethod as CreditCardData;
 
                 if (builder.TransactionModifier == TransactionModifier.EncryptedMobile) {
-                    et.SubElement(request, "mobile", card.MobileType);
                     et.SubElement(request, "token", card.Token);
+                    et.SubElement(request, "mobile", card.MobileType);
                 }
                 else {
                     var cardElement = et.SubElement(request, "card");
@@ -95,7 +95,7 @@ namespace GlobalPayments.Api.Gateways {
                     hash = GenerationUtils.GenerateHash(SharedSecret, timestamp, MerchantId, orderId, card.Number);
                 else {
                     if (builder.TransactionModifier == TransactionModifier.EncryptedMobile && card.MobileType == MobilePaymentMethodType.APPLEPAY)
-                        hash = GenerationUtils.GenerateHash(SharedSecret, timestamp, MerchantId, orderId, "", "", card.Token);
+                        hash = GenerationUtils.GenerateHash(SharedSecret, timestamp, MerchantId, orderId, builder.Amount.ToNumericCurrencyString(), builder.Currency, card.Token);
                     else if (builder.TransactionModifier == TransactionModifier.EncryptedMobile && card.MobileType == MobilePaymentMethodType.GOOGLEPAY)
                         hash = GenerationUtils.GenerateHash(SharedSecret, timestamp, MerchantId, orderId, builder.Amount.ToNumericCurrencyString(), builder.Currency, card.Token);
                     else
@@ -199,8 +199,18 @@ namespace GlobalPayments.Api.Gateways {
 
             #region FRAUD
             // fraud filter mode
-            if (builder.FraudFilterMode == FraudFilterMode.PASSIVE) {
-                et.SubElement(request, "fraudfilter").Set("mode", builder.FraudFilterMode.ToString());
+            if (builder.FraudFilterMode != FraudFilterMode.NONE) {
+                var fraudFilter = et.SubElement(request, "fraudfilter").Set("mode", builder.FraudFilterMode.ToString());
+                if(builder.FraudRules != null)
+                {
+                    var rules = et.SubElement(fraudFilter, "rules");
+                    foreach (var fraudRule in builder.FraudRules.Rules)
+                    {
+                        et.SubElement(rules, "rule")
+                            .Set("id",fraudRule.Key)
+                            .Set("mode", fraudRule.Mode.ToString());
+                    }
+                }
             }
 
             // recurring fraud filter
@@ -443,6 +453,13 @@ namespace GlobalPayments.Api.Gateways {
             }
             if (HostedPaymentConfig.FraudFilterMode != FraudFilterMode.NONE) {
                 request.Set("HPP_FRAUDFILTER_MODE", HostedPaymentConfig.FraudFilterMode.ToString());
+                if(HostedPaymentConfig.FraudFilterRules != null)
+                {
+                    foreach(var rule in HostedPaymentConfig.FraudFilterRules.Rules)
+                    {
+                        request.Set("HPP_FRAUDFILTER_RULE_"+rule.Key, rule.Mode.ToString());
+                    }
+                }
             }
             if (builder.RecurringType != null || builder.RecurringSequence != null) {
                 request.Set("RECURRING_TYPE", builder.RecurringType.ToString().ToLower());
@@ -748,14 +765,40 @@ namespace GlobalPayments.Api.Gateways {
                 },
                 AlternativePaymentResponse = new AlternativePaymentResponse
                 {
-                    BankAccount = root.Get("paymentmethoddetails")?.GetValue<string>("bankaccount"),
-                    AccountHolderName = root.Get("paymentmethoddetails")?.GetValue<string>("accountholdername"),
-                    Country = root.Get("paymentmethoddetails")?.GetValue<string>("country"),
-                    RedirectUrl = root.Get("paymentmethoddetails")?.GetValue<string>("redirecturl"),
-                    PaymentPurpose = root.Get("paymentmethoddetails")?.GetValue<string>("paymentpurpose"),
-                    PaymentMethod = root.Get("paymentmethoddetails")?.GetValue<string>("paymentmethod"),
-                }
+                    BankAccount = root.Has("paymentmethoddetails") ? root.Get("paymentmethoddetails")?.GetValue<string>("bankaccount"):null,
+                    AccountHolderName = root.Has("paymentmethoddetails") ? root.Get("paymentmethoddetails")?.GetValue<string>("accountholdername"): null,
+                    Country = root.Has("paymentmethoddetails") ? root.Get("paymentmethoddetails")?.GetValue<string>("country"):null,
+                    RedirectUrl = root.Has("paymentmethoddetails") ? root.Get("paymentmethoddetails")?.GetValue<string>("redirecturl"):null,
+                    PaymentPurpose = root.Has("paymentmethoddetails") ? root.Get("paymentmethoddetails")?.GetValue<string>("paymentpurpose"):null,
+                    PaymentMethod = root.Has("paymentmethoddetails") ? root.Get("paymentmethoddetails")?.GetValue<string>("paymentmethod"):null,
+                },
             };
+
+            // fraud response
+            if (root.Has("fraudresponse"))
+            {
+                Element fraudResponseElement = root.Get("fraudresponse");
+
+                FraudResponse fraudResponse = new FraudResponse
+                {
+                    mode = (FraudFilterMode) Enum.Parse(typeof(FraudFilterMode),fraudResponseElement.GetAttribute<string>("mode")),
+                    Result = fraudResponseElement.GetValue<string>("result"),
+                };
+
+                if (fraudResponseElement.Has("rules"))
+                {
+                    foreach (Element rule in fraudResponseElement.Get("rules").GetAll("rule"))
+                    {
+                        fraudResponse.Rules.Add(new FraudResponse.Rule
+                        {
+                            Id = rule.GetAttribute<string>("id"),
+                            Name = rule.GetAttribute<string>("name"),
+                            Action = rule.GetValue<string>("action"),
+                        });
+                    }
+                }
+                result.FraudResponse = fraudResponse;
+            }
 
             if (builder is ManagementBuilder mb && mb.MultiCapture) {
                 result.MultiCapturePaymentCount = mb.MultiCapturePaymentCount;